gpt4all-project/streamlit_app/requirements.txt
docker_py==1.10.6 
chromadb==0.5.0 
langchain==0.2.14
langchain-community==0.2.12
langchain-chroma==0.1.2
langchain-milus==0.1.3
selenium==4.22.0 
webdriver_manager==4.0.1 
pandas==2.2.2 
Pillow==10.3.0 
streamlit==1.36.0 